% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sym_pos_def.R
\name{is_sym_pos_def}
\alias{is_sym_pos_def}
\title{Check for positive definite matrix}
\usage{
is_sym_pos_def(A)
}
\arguments{
\item{A}{Numeric matrix (real or complex)}
}
\value{
TRUE indicates that the matrix A is symmetric (Hermitian) positive definite
}
\description{
The function checks the numeric matrix A for symmetricity (Hermitian for complex matrices)
first and then for non-zero eigenvalues
}
\examples{
A<-matrix(c(1,2,3,4),2,2)  # not symmetric
is_sym_pos_def(A)   # FALSE

A<-matrix(c(1,2,2,1),2,2)  # Symmetric but an eigen value is not positive
is_sym_pos_def(A)   # FALSE

A<-matrix(c(2,1,1,2),2,2)  # Both symmetric and eigen values are positive
is_sym_pos_def(A)   # TRUE

A<-matrix(c(4,2-3i,2+3i,3),2,2)  # Hermitian but one eigen value is not positive
is_sym_pos_def(A)   # FALSE
}
