% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeAC.R
\name{TimeAC}
\alias{TimeAC}
\title{Autocorrelation of a signal}
\usage{
TimeAC(x, biased = TRUE, demean = TRUE, normalize = FALSE, onesided = FALSE)
}
\arguments{
\item{x, }{Numeric, complex or logical vector of length \code{N}}

\item{biased}{logical TRUE (default) for biased estimator}

\item{demean}{logical to demean signal x before finding the ac (default TRUE)}

\item{normalize}{logical to return the normalized ac values (default)}

\item{onesided}{logical return the two-sided (default) or one-sided ac values}
}
\value{
A vector of autocorrelation values  of length \verb{2N-1} (includes positive and negative lags.)
}
\description{
This function returns the autocorrelation (ac) estimates of a signal (time series) x of length N.
Two types of estimators available : biased  (default) and unbiased. In addition,
this function has a plot utility to visualize the autocorrelation values with the default lag 50.
}
\details{

}
\examples{
TimeAC(c(1,2,3,4,5,6))   
TimeAC(rnorm(20))    
}
\seealso{
\code{\link[=TimeCC]{TimeCC()}} .
}
