% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeCC.R
\name{TimeCC}
\alias{TimeCC}
\title{Crosscorrelation of a signal}
\usage{
TimeCC(x, y)
}
\arguments{
\item{xy}{Numeric, complex or logical vectors that indicate signals of length \code{N} each}

\item{biased}{logical TRUE (default) for biased estimator}

\item{demean}{logical to demean signal x before finding the ac (default TRUE)}

\item{normalize}{logical to return the normalized ac values (default)}

\item{onesided}{logical return the two-sided (default) or one-sided ac values}
}
\value{
A vector of crosscorrelation values  of length \verb{2N-1} (includes positive and negative lags.)
}
\description{
This function returns the crosscorrelation (cc) estimates of a signal (time series) x of length N.
Two types of estimators available : biased  (default) and unbiased. In addition,
this function has a plot utility to visualize the autocorrelation values with the default lag 50.
}
\examples{
TimeCC(c(1,2,3,4,5,6),c(6,5,4,3,2,1))   
TimeCC(rnorm(20),rnorm(20))    
}
\seealso{
\code{\link[=TimeAC]{TimeAC()}} .
}
