% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_sc.R
\name{var_sc}
\alias{var_sc}
\title{Variance of Pearson Correlation}
\usage{
var_sc(x, y)
}
\arguments{
\item{xy}{Numeric vectors that indicate signals of length \code{N}}
}
\value{
scalar of variance
}
\description{
This function returns the variance of the Pearson correlation between two signals
}
\examples{
var_sc(c(1,2,3,4,5,6),c(6,5,4,3,2,1))   
}
